+++ src/terminal.h
@@ -5,7 +5,7 @@
 
 typedef struct term_s {
   struct termios oterm;
-  i32 cols, rows;
+  usize cols, rows;
 } term_t;
 
 void term_init(void);
